Added extra functionality to TurnBased, added different abilities

+++ TurnBased/src/turnBasedCharacters/Enemy.java
@@ -2,7 +2,7 @@ package turnBasedCharacters;
 
 public class Enemy extends GameCharacter
 {
-	public Enemy(String name, int health, int hitChance, int minDmg, int maxDmg)
+	public Enemy(String name, int health, double hitChance, int minDmg, int maxDmg)
 	{
 		super(name, health, hitChance, minDmg, maxDmg);
 		this.setType("Enemy");
